requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pywinrm==0.5.0
wakeonlan==3.1.0